Reference

MCDTS.AbstractTreeElement — Type

The MCDTS algorithm is implemented as a tree with different kind types encoding
the leafs and the root of the tree. AbstractTreeElement is the abstract type of
these types.

MCDTS.CCM_ρ — Type

CCM_ρ <: AbstractLoss

Constructor for the CCM_ρ loss function (correlation coefficient of the
convergent cross mapping) based on Sugihara et al.[^Sugihara2012], see also
[`ccm`](@ref). In this case MCDTS tries to maximize the correlation coefficient
of the convergent cross mapping from the input `data` and `Y_CCM`, the time
series CCM should cross map to from `data` (see [`mcdts_embedding`](@ref)).

## Fieldnames
* `timeseries`: The time series CCM should cross map to.
* `threshold::Float`: A threshold for the sufficient correlation of the
  cross-mapped values and the true values from `Y_CMM` for the current embedding.
  When the correlation coefficient exeeds this threshold in an embedding cycle
  the embedding stops.

## Defaults
* When calling `CCM_ρ(timeseries)`, a CCM_ρ-object is created, storing
  `timeseries`, which is considered to be causally depended and the
  `threshold=1` is used, i.e. no threshold, since the correlation coefficient
  can not exceed 1.

[^Sugihara2012]: Sugihara et al., [Detecting Causality in Complex Ecosystems. Science 338, 6106, 496-500](https://doi.org/10.1126/science.1227079)

MCDTS.Continuity_function — Type

Continuity_function <: AbstractDelayPreselection

Constructor for the continuity function `⟨ε★⟩` by Pecora et al.[^Pecora2007],
see [`pecora`](@ref).

## Fieldnames
* `K::Int`: the amount of nearest neighbors in the δ-ball. Must be at
  least 8 (in order to gurantee a valid statistic). `⟨ε★⟩` is computed taking
  the minimum result over all `k ∈ K` (read algorithm description in [`pecora`](@ref)).
* `samplesize::Real`: determine the fraction of all phase space points
  to be considered (fiducial points v) to average ε★ to produce `⟨ε★⟩`.
* `α::Real = 0.05`: The significance level for obtaining the continuity statistic
* `p::Real = 0.5`: The p-parameter for the binomial distribution used for the
  computation of the continuity statistic ⟨ε★⟩.

## Defaults
* When calling `Continuity_function()` a Continuity_function-object is constructed
  with `K=13`, `samplesize=1.`, `α=0.05` and `p=0.5`.

MCDTS.FNN_statistic — Type

FNN_statistic <: AbstractLoss

Constructor for the FNN-statistic loss function (false nearest neighbor) based
on Hegger & Kantz [^Hegger1999].

## Fieldnames
* `threshold::Float`: A threshold for the tolerable cumulative FNN decrease
  for the current embedding. When the fraction of FNNs fall below this threshold
  in an embedding cycle the embedding stops.
* `r::Float`: The FNN-distance-expansion threshold (typically set to 2).

## Defaults
* When calling `FNN_statistic()`, a FNN_statistic-object is created, which uses no
  threshold and uses the FNN-inter threshold `r=2`.
* When calling `FNN_statistic(threshold)`, a FNN_statistic-object is created, which uses
  the given `threshold` and uses the FNN-inter threshold `r=2`.

[^Hegger1999]: Hegger & Kantz, [Improved false nearest neighbor method to detect determinism in time series data. Physical Review E 60, 4970](https://doi.org/10.1103/PhysRevE.60.4970).

MCDTS.L_statistic — Type

L_statistic <: AbstractLoss

Constructor for the L-statistic loss function based on Uzal et al.[^Uzal2011]. Here
we consider the decrease of the L-statistic `ΔL` in between embedding cycles,
according to Kraemer et al.[^Kraemer2021] ([`pecuzal_embedding`](@ref)).

## Fieldnames
* `threshold::Float`: A threshold for the tolerable `ΔL` decrease for the current
  embedding. When `ΔL` exceeds this threshold in an embedding cycle the embedding
  stops. Note that `ΔL` is a negative value therefore `threshold` must be a small
  negative number.
* `KNN::Int`: the amount of nearest neighbors considered, in order to compute the
 L-statistic, in particular `σ_k^2` (read algorithm description [`uzal_cost`]@ref).
* `tws::AbstractRange{Int}`: Customization of the sampling of the different time horizons
  (T's), when computing Uzal's L-statistics. Here any kind of integer ranges (starting at 2)
  are allowed.

## Defaults
* When calling `L_statistic()`, a L_statistic-object is created, which uses no
  threshold and consideres 3 nearest neighbors for time horizons `tws=2:100`.
* When calling `L_statistic(threshold)`, a L_statistic-object is created, which uses
  the given `threshold` and consideres 3 nearest neighbors for time horizons `tws=2:100`.
* When calling `L_statistic(threshold,KNN)`, a L_statistic-object is created, which uses
  the given `threshold`, consideres `KNN` nearest neighbors for time horizons `tws=2:100`.

[^Kraemer2021]: Kraemer, K.H., Datseris, G., Kurths, J., Kiss, I.Z., Ocampo-Espindola, Marwan, N. (2021). [A unified and automated approach to attractor reconstruction. New Journal of Physics 23(3), 033017](https://iopscience.iop.org/article/10.1088/1367-2630/abe336).

[^Uzal2011]: Uzal, L. C., Grinblat, G. L., Verdes, P. F. (2011). [Optimal reconstruction of dynamical systems: A noise amplification approach. Physical Review E 84, 016223](https://doi.org/10.1103/PhysRevE.84.016223).

MCDTS.MCDTSOptimGoal — Type

MCDTSOptimGoal <: AbstractMCDTSOptimGoal

Constructor, which handles the loss-/objective function `Γ` and the delay
pre-selection statistic `Λ` MCDTS uses.

## Fieldnames
* `Γ::AbstractLoss`: Chosen loss-function, see the so far available
  [`L_statistic`](@ref), [`FNN_statistic`](@ref), [`CCM_ρ`](@ref) and
  [`Prediction_error`](@ref).
* `Λ::AbstractDelayPreselection`: Chosen delay Pre-selection method, see the so
  far available [`Continuity_function`](@ref) and [`Range_function`](@ref).

## Defaults
* TBD

MCDTS.MCDTSpredictionType — Type

MCDTSpredictionType <: AbstractMCDTSpredictionType

Constructor, which determines the way how predictions are made technically.

## Fieldnames
* `loss::AbstractPredictionLoss`: Indicates the way of computing the prediction error.
   See [`PredictionLoss`](@ref) for information on how to construct this object.
* `method::AbstractPredictionMethod`: The method based on the state space reconstruction,
   which makes the actual prediction. See [`local_model`](@ref)

## Default settings
* When calling `MCDTSpredictionType()` a MCDTSpredictionType-object is constructed
  with a `local_zeroth`-predictor [`local_model`](@ref), using 2 nearest neighbors
  and a 1-step-ahead-prediction. The loss-function is the root mean squared prediction
  error over all components [`PredictionLoss`](@ref).

MCDTS.Node — Type

mutable struct Node{T}

A node of the tree. Each node contains its children and information about the current embedding.

## Fieldnames:
* `τ::Int`: The delay value of this node
* `L::T`: The value of the cumulative ΔL statistic at this node
* `τs::Array{Int,1}`: The complete vector with all τs chosen along this path up until this node
* `ts::Array{Int,1}`: The complex vector which of the possibly multivariate time series is used at each embedding step i
* `children::Union{Array{Node,1},Nothing}`: The children of this node
* `temp::S`: additional "work"/"temporary" array/field that can be manipulated and saved by the optimization e.g. to reuse prior computations

MCDTS.PredictionLoss — Type

PredictionLoss <: AbstractPredictionLoss

Constructor, which indicates the way of computing the prediction error. This
object is used for the constructor, which determines the way how predictions are
made methodologically [`MCDTSpredictionType`](@ref).

## Fieldnames
* `type::Int` is an integer, which encodes the type of prediction error:
* For `type = 1` the root mean squared prediction error over the first component,
  i.e. the timeseries, which needs to be predicted, is used. (default)
* For `type = 2` the root mean squared prediction error over all components
  (dimensionality of the state space) is used.
* For `type = 3` the mean Kullback-Leibler Distance of the predicted and the true
  values of the first component, i.e. the timeseries, which needs to be predicted, is used.
* For `type = 4` the mean Kullback-Leibler Distance of the predicted and the true
  values over all components (dimensionality of the state space) is used.

## Default settings
* When calling `PredictionLoss()` a PredictionLoss-object is constructed with
  fieldname `type = 1` (≡root mean squared prediction error over all components)

MCDTS.Prediction_error — Type

Prediction_error <: AbstractLoss

Constructor for the Prediction_error loss function.

## Fieldnames
* `PredictionType::MCDTSpredictionType`: Determines the prediction type by
  setting a prediction-method and the way the prediction error is measured,
  see [`MCDTSpredictionType`](@ref).
* `threshold::Float`: A threshold for the sufficient minimum prediction error
  for the current embedding. When the prediction error, specified in
  `PredictionType`, falls below this threshold in an embedding cycle the
  embedding stops.

## Defaults
* When calling `Prediction_error()`, a Prediction_error-object is created,
  which uses the threshold 0, i.e. no threshold and a zeroth-order predictor
  (see [`MCDTSpredictionType`](@ref), [`PredictionLoss`](@ref) &
  [`local_model`](@ref))

MCDTS.Range_function — Type

Range_function <: AbstractDelayPreselection

Constructor for a range of possible delay values. In this case there is
actually no "pre-selection" of delay, but rather all possible delays, given
in the input `τs` (see, [`mcdts_embedding`](@ref)) are considered. This can significantly affect the
computation time. There are no fieldnames, simply construct by typing
`RangeFunction()`.

MCDTS.Root — Type

mutable struct Root <: AbstractTreeElement

The 'start'/root of Tree. Each node contains its children. The root contains the starting branches/nodes.
For initialization type `r = Root()`.

## Fieldnames:
* `children::Union{Array{Node,1},Nothing}`: The first nodes of the tree.
* `Lmin`; Is the global minimum of the cumulative ΔL statistic found so far.

MCDTS.local_model — Type

local_model <: AbstractPredictionMethod

Constructor, which indicates the local state space prediction model.

## Fieldnames
* `method::String`: Could be `"zeroth"` (averaged `Tw`-step-ahead image of the
 `KNN`-nearest neighbors) or `"linear"` (local linear regression on the
 `KNN`-nearest neighbors).
* `KNN::Int`: The number of considered nearest neighbors.
* `Tw::Int` : The prediction horizon in sampling units.

## Default settings
* When calling `local_model()` a local_model-object is constructed with a zeroth
  order prediction scheme, 2 nearest neighbors and a 1-step-ahead prediction.
* When calling `local_model(method)` a local_model-object is constructed with a
  `method`-prediction scheme, 2 nearest neighbors and a 1-step-ahead prediction.
* When calling `local_model(method,KNN)` a local_model-object is constructed with a
 `method`-prediction scheme, `KNN` nearest neighbors and a 1-step-ahead prediction.

MCDTS.all_neighbors — Function

all_neighbors(A::Dataset, stype, w = 0) → idxs, dists

Find the neighbors of all points in A using search type stype (either NeighborNumber or WithinRange) and w the Theiler window.

This function is nothing more than a convinience call to Neighborhood.bulksearch.

MCDTS.ar_process — Method

Generate data from a AR(1) process for a initial value `u0`, a AR-coefficient
`α` and a white noise scaling parameter `p`. Return a time series of length `N`.

MCDTS.backprop! — Method

backprop!(n::Root,τs,ts,L_min)

Backpropagation of the tree spanned by all children in `n` (for this run).
All children-nodes L-values get set to the final value achieved in this run.
This function is ususally called be [`expand!`](@ref).

MCDTS.best_embedding — Method

best_embedding(r::Root)

Given the root `r` of a tree, return the best embedding in the form of the
final node at the end of the best embedding.

MCDTS.ccm — Method

ccm(X, y; kwargs...) → ρ, y_hat

Compute the convergent crossmapping (CCM) (Sugihara et al. 2012) of a
vector time series `X` (an embedded time series `x`) on the time series `y`
NOTE: 'X' and 'y' must have the same length and you have to make sure that
'y' starts at the same time index as 'X' does. - When using [`genembed`](@ref)
with negative delays to construct `X` from `x`, which is mandatory here, then
'y' needs to be shifted by the largest negative delay value, which has been
used to construct `X`.

Returns the correlation coefficient of `y` and its predicted values for `y_hat`,
based on the nearest neighbour structure of `X`.
It is said that 'y' causes 'x', if ρ increases with increasing time series
length AND ρ is "quite high".

Keyword arguments:
*`metric = Euclidean()`: The metric for vector distance computation.
*`w::Int = 1`: The Theiler window in sampling units.
*`lags::Array = [0]`: The lag for the cross mapping, in order to detect time lagged
                      causal relationships. The output ρ is an array of size
                      `length(lags)`, the output Y_hat is the one corresponding
                      to a lag of zero.

MCDTS.choose_children — Method

choose_children(n::AbstractTreeElement, τ::Int, t:Int)

Pick one of the children of the tree node n with values τ and t. If there is none, return nothing.

MCDTS.choose_next_node — Function

choose_next_node(n::Union{Node,Root}, func, Lmin_global, i_trial::Int=1)

Returns one of the children of based on the function `func(Ls)->i_node`,
Lmin_global is the best L value so far in the optimization process, if any of
the input Ls to choose from is smaller than it, it is always chosen.
`choose_mode` is only relevant for the first embedding step right now: it
determines if the first step is chosen uniform (`choose_mode==0`) or with the
`func` (`choose_mode==1`).

MCDTS.columns — Function

columns(dataset) -> x, y, z, ...

Return the individual columns of the dataset.

MCDTS.compute_KL_divergence — Method

Compute the Kullback-Leibler-Divergence of the two Vectors `a` and `b`.

MCDTS.compute_abs_err — Method

Compute the total absolute error between `prediction` and `reference`

MCDTS.compute_costs_from_prediction — Method

Compute the in-sample prediction costs based on the loss-metric determined
by PredictionLoss

MCDTS.compute_delta_L — Method

compute_delta_L(s, τs, (js,) T_max; KNN = 3, w = 1, metric = Euclidean) → ΔL

Compute the overall L-decrease `ΔL` of a given embedding of the time series
`s::Vector` with the delay values `τs` up to a maximum `T`-value `T_max`. We
respect the Theiler window `w`, the chosen `metric` and the number of considered
nearest neighbors `KNN`. It is also possible to compute `ΔL` for a multivariate
input `Y::Dataset`. Then one additionally needs to supply a vector `js`, which
lists the chosen time series corresponding to the given delay values in `τs`.
This is similar to the procedure in [`genembed`]@ref. The computations are based
on z-standardized input for ensuring comparability.

MCDTS.compute_loss — Function

Compute the loss of a given delay-preselection statistic `dps` and the loss
determined by `optimalg.Γ`.

MCDTS.compute_loss — Method

Return the loss based on the negative correlation coefficient for CCM.

MCDTS.compute_loss — Method

Return the loss based on the FNN-statistic `FNN` and indices `max_idx`  for all local maxima in dps

MCDTS.compute_loss — Method

Return the loss based on the maximum decrease of the L-statistic `L_decrease` and corresponding
delay-indices `max_idx` for all local maxima in ε★

MCDTS.compute_loss — Method

Return the loss based on a `Tw`-step-ahead local-prediction.

MCDTS.compute_mse — Method

Compute the mean squared error between `prediction` and `reference`

MCDTS.difference_distribution — Method

calculate σ(ζ) = ∫ρ(x) ρ(ζ-x) dx , approximating ζ to be in bin start (whole bin counts).

MCDTS.embed_for_prediction — Method

embed_for_prediction(Y::Dataset, x::Vector, τ::Int) → Y_embed

Embeds the trajectory (or Vector) `Y` in an additional dimension, using the time
series `x` and the provided lag `τ`. Here we enforce a causal embedding, meaning
that we shift `x` by the negative values `τ`.

MCDTS.embedding_cycle — Method

Perform a potential embedding cycle from the multi- or univariate Dataset `Ys`.
Return the possible delays `τ_pot`, the associated time series `ts_pot` and
the corresponding L-statistic-values, `L_pot` for each peak, i.e. for each
(`τ_pot`, `ts_pot`) pair. If `FNN=true`, `L_pot` stores the corresponding
fnn-statistic-values.

MCDTS.expand! — Method

expand!(n::Union{Node,Root}, optimalg::AbstractMCDTSOptimGoal, data::Dataset,
                                        w::Int, delays, choose_mode; kwargs...)

This is one single rollout and backprop of the tree. For details please see
the accompanying paper [^Kraemer2021b].

* `n`: Starting node
* `optimalg::AbstractMCDTSOptimGoal`: Determines the delay preselection and
  cost function (see [`MCDTSOptimGoal`](@ref)).
* `data`: data
* `w`: Theiler Window
* `delays = 0:100`: The possible time lags
* `choose_mode::Int=0`: Possibility for different modes of choosing the next
  node based on which trial this is.

## Keyword arguments
* See [`mcdts_embedding`](@ref) for a list of all keywords.

MCDTS.fnn_embedding_cycle — Function

fnn_embedding_cycle(NNdist, NNdistnew, r=2) -> FNNs

Compute the amount of false nearest neighbors `FNNs`, when adding another component
to a given (vector-) time series. This new component is the `τ`-lagged version
of a univariate time series. `NNdist` is storing the distances of the nearest
neighbor for all considered fiducial points and `NNdistnew` is storing the
distances of the nearest neighbor for each fiducial point in one embedding
dimension higher using a given `τ`. The obligatory threshold `r` is by default
set to 2.

MCDTS.genembed_for_prediction — Method

genembed_for_prediction(Y, τs::Vector (,ts::Vector)) → Y_embed

Embeds the trajectory (or Vector) `Y` using the provided lags `τ`. If `Y` is a
dataset, a vector `ts` must be provided, which stores the indices of the time
series contained in `Y`, which needs to be used for each embedding cycle. Here
we enforce a causal embedding, meaning that we shift the time series `Y` (or any
of the time series in `Y`, if `Y` is a Dataset) by the negative values `τ` from
`τs`.

MCDTS.get_ar_prediction — Method

get_ar_prediction(x::Vector, coeffs::Vector; kwargs...) → Y_predict

Computes a prediction `Y_predict` of the AR-model determined by the coefficients
in `coeffs`. The order of the AR-model equals the length of `coeffs`. `x` can be
a long vector (the time series), but needs to contain at least `length(coeffs)`
values, in order to initialize the model. If the time horizon `Tw` is larger than
1, an iterated one-step prediction is peformed.

Keywords:
* `Tw::Int = 1`: Time horizon for the prediction
* `c::Real = 0`: Offset-parameter for AR-model
* `rng::AbstractRNG = Random.GLOBAL_RNG`: Random number generator

MCDTS.get_binomial_table — Method

get_binomial_table(p, α; trial_range::Int=8) -> `δ_to_ε_amount`, Dict(δ_points => ϵ_points)

compute the numbers of points from the δ-neighborhood, which need to fall outside the ϵ-neighborhood, in order to reject the Null Hypothesis at a significance level α. One parameter of the binomial distribution is p, the other one would be the number of trials, i.e. the considered number of points of the δ-neighborhood. trial_range determines the number of considered δ-neighborhood-points, always starting from 8. For instance, if trial_range=8 (Default), then δ-neighborhood sizes from 8 up to 15 are considered. Return δ_to_ε_amount, a dictionary with δ_points as keys and the corresponding number of points in order to reject the Null, ϵ_points, constitute the values.

MCDTS.get_delay_statistic — Method

get_delay_statistic(optimalg.Λ<: AbstractDelayPreselection, Ys, τs, w, τ_vals, ts_vals; kwargs... )

Compute the delay statistic according to the chosen method in `optimalg.Λ` (see [`MCDTSOptimGoal`](@ref))

MCDTS.get_embedding_params_according_to_loss — Method

get_embedding_params_according_to_loss(Γ::AbstractLoss, τ_pot, ts_popt, L_pot, L_old)

Helper function for [`get_potential_delays`](@ref). Computes the potential
delay-, time series- and according Loss-values with respect to the actual loss
in the current embedding cycle.

MCDTS.get_max_idx — Method

get_max_idx(Λ::AbstractDelayPreselection, dps::Vector, τ_vals, ts_vals) → max_idx

Compute the candidate delay values from the given delay pre-selection statistic
`dps` with respect to `Λ`, which determined how `dps` was obtained and how
to select the candidates (e.g. pick the maxima of `dps` in case of the
`Λ` being the Continuity function). See [`Continuity_function`](@ref) and
[`Range_function`](@ref).

MCDTS.get_maxima — Method

Return the local maxima of the given time series s and its indices

MCDTS.get_potential_delays — Method

get_potential_delays(optimalg::AbstractMCDTSOptimGoal, Ys::Dataset, τs, w::Int, τ_vals,
                ts_vals, L_old ; kwargs...]) → τ_pot, ts_pot, L_pot, flag

Compute the potential delay `τ_pot` and time series values `ts_pot`, which would
each result in a potential Loss-statistic value `L_pot`, by using an
embedding method specified in `optimalg` [^Kraemer2021b] (see [`MCDTSOptimGoal`](@ref))
and for a range of possible delay values `τs`. The input dataset `Ys` can be
multivariate. `w` is the Theiler window (neighbors in time with index `w` close
to the point, that are excluded from being true neighbors. `w=0` means to
exclude only the point itself, and no temporal neighbors. In case of multivariate
time series input choose `w` as the maximum of all `wᵢ's`. `τ_vals` and `ts_vals`
describe the embedding up to the current embedding cycle.

## Keyword arguments
* See [`mcdts_embedding`](@ref) for a list of all keywords.

MCDTS.integral_σ — Method

approximately calculate ∫_0 ^(ξ) σ(ζ) dζ.

MCDTS.iterated_local_linear_prediction — Method

iterated_local_linear_prediction(Y::Dataset, K::Int = 5, Tw::Int = 2; kwargs...) → Y_predict

Perform an iterated one step forecast over `Tw` time steps using the local linear
prediction algorithm. `Y_predict` is a Dataset of length `Tw` and dimension like
`Y`.

Keywords:
* `metric = Euclidean()`: Metric used for distance computation
* `theiler::Int = 1`: Theiler window for excluding serially correlated points from
   the nearest neighbour search.
* `verbose::Bool = false`: When set to `true`, the function prints the actual time
  step, which it is computing.

MCDTS.iterated_local_linear_prediction_embed — Method

iterated_local_linear_prediction_embed(Y::Dataset, τs::Vector, K::Int = 5, Tw::Int = 2; kwargs...) → Y_predict
Perform an iterated one step forecast over `Tw` time steps using the local linear
prediction algorithm. `Y_predict` is a Dataset of length `Tw` and dimension like
`Y`. In contrast to `iterated_local_linear_prediction()` we here use the time
delays `τs` to reconstruct all components of a predicted trajectory point from
the 1st component, which is obtained from the local model. This only works for
univariate embeddings.

Keywords:
* `metric = Euclidean()`: Metric used for distance computation
* `theiler::Int = 1`: Theiler window for excluding serially correlated points from
   the nearest neighbour search.
* `verbose::Bool = false`: When set to `true`, the function prints the actual time
  step, which it is computing.

MCDTS.iterated_local_zeroth_prediction — Method

iterated_local_zeroth_prediction(Y::Dataset, K::Int = 5, Tw::Int = 2; kwargs...) → Y_predict

Perform an iterated one step forecast over Tw time steps using the local zeroth prediction algorithm. Y_predict is a Dataset of length Tw and dimension like Y.

Keywords:

metric = Euclidean(): Metric used for distance computation

theiler::Int = 1: Theiler window for excluding serially correlated points from the nearest neighbour search.

verbose::Bool = false: When set to true, the function prints the actual time step, which it is computing.

MCDTS.iterated_local_zeroth_prediction_embed — Method

iterated_local_zeroth_prediction_embed(Y::Dataset, τs::Vector, K::Int = 5, Tw::Int = 2; kwargs...) → Y_predict

Perform an iterated one step forecast over Tw time steps using the local linear prediction algorithm. Y_predict is a Dataset of length Tw and dimension like Y. In contrast to iterated_local_linear_prediction() we here use the time delays τs to reconstruct all components of a predicted trajectory point from the 1st component, which is obtained from the local model. This only works for univariate embeddings.

Keywords:

metric = Euclidean(): Metric used for distance computation

theiler::Int = 1: Theiler window for excluding serially correlated points from the nearest neighbour search.

verbose::Bool = false: When set to true, the function prints the actual time step, which it is computing.

MCDTS.jrp_rr_frac — Method

Computes the similarity between recurrence plots `RP₁` and `RP₂`. Outputs the
fraction of recurrences rates gained from RP₁ and of the joint recurrence
plot `RP₁ .* RP₂`.

MCDTS.linear_prediction_cost — Method

linear_prediction_cost(Y::Dataset; kwargs...) → Cost

Compute the mean squared one-step prediction error Cost of the Dataset Y. The prediction is based on local_linear_prediction.

Keyword arguments

samplesize = 1.0: Number of considered fiducial points v as a fraction of input state space trajectory Y's length, in order to average the conditional variances and neighborhood sizes (read algorithm description) to produce L.

K = 3: the amount of nearest neighbors considered, in order to compute the mean squared prediction error (read algorithm description).

metric = Euclidean(): metric used for finding nearest neigbhors in the input state space trajectory `Y.

w = 1: Theiler window (neighbors in time with index w close to the point, that are excluded from being true neighbors). w=0 means to exclude only the point itself, and no temporal neighbors.

Tw = 1: The time horizon for predictions. The Cost is the average error over these timesteps.

MCDTS.linear_prediction_cost_KL — Method

linear_prediction_cost_KL(Y::Dataset; kwargs...) → Cost

Compute the KL-divergence Cost of the Dataset Y. The prediction is based on local_linear_prediction.

Keyword arguments

samplesize = 1.0: Number of considered fiducial points v as a fraction of input state space trajectory Y's length, in order to average the conditional variances and neighborhood sizes (read algorithm description) to produce L.

K = 3: the amount of nearest neighbors considered, in order to compute the mean squared prediction error (read algorithm description).

metric = Euclidean(): metric used for finding nearest neigbhors in the input state space trajectory `Y.

w = 1: Theiler window (neighbors in time with index w close to the point, that are excluded from being true neighbors). w=0 means to exclude only the point itself, and no temporal neighbors.

Tw = 1: The time horizon for predictions. The Cost is the average error over these timesteps.

MCDTS.local_linear_prediction — Method

local_linear_prediction(Y::Dataset, K::Int = 5; kwargs...) → x_pred, e_expect

Perform a prediction for the time horizon `Tw` (default = 1) by a locally linear
fit. Based on `K` nearest neighbours of the last point of the given trajectory
`Y`, we fit a linear model to these points and their `Tw`-step ahead images. The
output `x_pred` is, thus, the `Tw`-step ahead prediction vector.
The function also returns `e_expect`, the expected error on the prediction `x_pred`,
computed as the mean of the RMS-errors of all `K`-neighbours-errors.

Keywords:
* `metric = Euclidean()`: Metric used for distance computation
* `theiler::Int = 1`: Theiler window for excluding serially correlated points from
   the nearest neighbour search.
* `Tw::Int = 1`: The prediction time in sampling units. If `Tw > 1`, a multi-step
  prediction is performed.

MCDTS.local_random_analogue_prediction — Method

local_random_analogue_prediction(Y::Dataset, K::Int; kwargs...) → Y_predict

Compute a one step ahead prediction Y_predict of the input Y, based on K nearest neighbors. Here the prediction is a random pick from all K-nearest neighbour images and, thus, invokes some kind of randomness.

Keywords:

metric = Euclidean(): Metric used for distance computation

theiler::Int = 1: Theiler window for excluding serially correlated points from the nearest neighbour search.

MCDTS.local_zeroth_prediction — Method

local_zeroth_prediction(Y::Dataset, K::Int = 5; kwargs...) → x_pred, e_expect

Perform a "zeroth" order prediction for the time horizon Tw (default = 1). Based on K nearest neighbours of the last point of the given trajectory Y, the Tw-step ahead prediction is simply the mean of the images of these K-nearest neighbours. The output x_pred is, thus, the Tw-step ahead prediction vector. The function also returns e_expect, the expected error on the prediction x_pred, computed as the mean of the RMS-errors of all K-neighbours-errors.

Keywords:

metric = Euclidean(): Metric used for distance computation

theiler::Int = 1: Theiler window for excluding serially correlated points from the nearest neighbour search.

Tw::Int = 1: The prediction time in sampling units. If Tw > 1, a multi-step prediction is performed.

MCDTS.make_prediction — Method

make_prediction(pred_meth::AbstractPredictionMethod, Y::AbstractDataset{D, ET};
        K::Int = 3, w::Int = 1, Tw::Int = 1, metric = Euclidean()) → prediction

Compute a in-sample `Tw`-time-steps-ahead of the data `Y`, using the prediction
method `pred_meth`. `w` is the Theiler window and `K` the nearest neighbors used.

MCDTS.mc_delay — Method

Legacy name, please use [`mcdts_embedding`](@ref)

MCDTS.mcdts_embedding — Method

mcdts_embedding

## Convenience / default option call

mcdts_embedding(data::Dataset, N::Int=100; kwargs...)

Do the MCDTS embedding of the `data` with `N` trials according to the
PECUZAL algorithm [^Kraemer2021], returns the tree. Embedding parameters
like the Theiler window, considered delays and the function that chooses the
next embedding step are all estimated automatically or the default option is
used. `data` is a `DynamicalSystems.Dataset`.

## All options

mcdts_embedding(data::DynamicalSystems.Dataset, optimalg::AbstractMCDTSOptimGoal,
                        w::Int, delays::AbstractRange{D}, N::Int=40;  kwargs...)

* `optimalg::AbstractMCDTSOptimGoal` determines how the embedding is performed in
  each cycle. Specifically it sets the delay pre-selection statistic Λ, which
  pre-selects the potential delays in each embedding cycle as well as the the
  Loss-Statistic Γ, which determines the Loss to be minimized by MCDTS
  [^Kraemer2021b] (see [`MCDTSOptimGoal`](@ref)).
* `w` is the Theiler window (neighbors in time with index `w` close to the point,
  that are excluded from being true neighbors. `w=0` means to exclude only the
  point itself, and no temporal neighbors. In case of multivariate time series
  input choose `w` as the maximum of all `wᵢ's`. As a default in the convience
  call this is estimated with a mutual information minimum method of DelayEmbeddings.jl
* `delays = 0:100`: The possible time lags
* `N::Int`: The number of tree expansions

## Keyword Arguments
* `choose_func`: Function to choose next node in the tree with, default
  choice: `(L)->(MCDTS.softmaxL(L,β=2.))`
* `max_depth = 20`: Threshold, which determines the algorithm. It either breaks,
  when it converges, i.e. when there is no way to reduce the cost-function any
  further, or when this threshold is reached.
* `verbose`: Either `true` or `false` (default); prints status of embedding optimization.
* `metric`: norm for distance computation (default is `Euclidean()`)

* `KNN = 3`: The number of nearest neighbors considered in the computation of
  the L-statistic.
* `FNN:Bool = false`: Determines whether the algorithm should minimize the
  L-statistic or the FNN-statistic.
* `PRED::Bool = false`: Determines whether the algorithm should minimize the
  L-statistic or a cost function based on minimizing the `Tw`-step-prediction error
* `Tw::Int = 1`: If `PRED = true`, this is the considered prediction horizon
* `linear::Bool=false`: If `PRED = true`, this determines whether the prediction shall
  be made on the zeroth or a linear predictor.
* `PRED_mean::Bool=false`: If `PRED = true`, this determines whether the prediction shall
  be optimized on the mean MSE of all components or only on the 1st-component (Default)
* `PRED_L::Bool=false`: If `PRED = true`, this determines whether the prediction shall
  be optimized on possible delay values gained from the continuity statistic or on
  delays = 0:25 (Default)
* `PRED_KL::Bool=false`: If `PRED = true`, this determines whether the prediction shall
  be optimized on the Kullback-Leibler-divergence of the in-sample prediction and
  the true in-sample values, or if the optimization shall be made on the MSE of them (Default)
* `CCM:Bool=false`: Determines whether the algorithm should maximize the CCM-correlation
  coefficient of the embedded vector from `Ys` and the given time series `Y_CCM`.
* `Y_CCM`: The time series CCM should cross map to.
* `threshold::Real = 0`: The algorithm does not pick a peak from the continuity
  statistic, when its corresponding `ΔL`/FNN-value exceeds this threshold. Please
  provide a positive number for both, `L` and `FNN`-statistic option (since the
  `ΔL`-values are negative numbers for meaningful embedding cycles, this threshold
  gets internally sign-switched).
* `tws::Range = 2:delays[end]`: Customization of the sampling of the different T's,
  when computing Uzal's L-statistics. Here any kind of integer ranges (starting
  at 2) are allowed, up to `delays[end]`.

  [^Kraemer2021]: Kraemer, K.H., Datseris, G., Kurths, J., Kiss, I.Z., Ocampo-Espindola, Marwan, N. (2021). [A unified and automated approach to attractor reconstruction. New Journal of Physics 23(3), 033017](https://iopscience.iop.org/article/10.1088/1367-2630/abe336).
  [^Kraemer2021b]: Kraemer, K.H., Gelbrecht, M., Pavithran, I., Sujith, R. I. and Marwan, N. (2021). [Optimal state space reconstruction via Monte Carlo Decision Tree Search. Submitted to Nonlinear Dynamics](https://doi.org/10.21203/rs.3.rs-899760/v1)

MCDTS.mfnn — Method

Computes the mututal false nearest neighbours (mfnn) for a reference trajectory
`Y_ref` and a reconstruction `Y_rec` after [^Rulkov1995].

## Keyword arguments
*`w = 1`: Theiler window for the surpression of serially correlated neighbors in
    the nearest neighbor-search
*`kNN = 1`: The number of considered nearest neighbours (in the paper always 1)

[^Rulkov1995]: Rulkov, Nikolai F. and Sushchik, Mikhail M. and Tsimring, Lev S. and Abarbanel, Henry D.I. (1995). [Generalized synchronization of chaos in directionally coupled chaotic systems. Physical Review E 51, 980](https://doi.org/10.1103/PhysRevE.51.980).

MCDTS.moving_average — Method

Moving average of a timeseries `vs` over a window `n`

MCDTS.next_embedding — Method

next_embedding(n::Node, optimalg::AbstractMCDTSOptimGoal, Ys::Dataset{D, T},
                        w::Int, τs; kwargs...) → τ_pot, ts_pot, L_pot, flag

Performs the next embedding step. For the actual embedding contained in tree
leaf `n` compute as many statistics determined by `optimalg.Λ`
(see [`MCDTSOptimGoal`](@ref)) as there are time series in the Dataset
`Ys` for a range of possible delays `τs`. Return the values for the best delay
`τ_pot`, its corresponding time series index `ts_pot` the according Loss-value
`L_pot` and `flag`, following the minimization of the Loss determined by
`optimalg.Γ`.

## Keyword arguments
* See [`mcdts_embedding`](@ref) for a list of all keywords.

## Returns
* `τ_pot`: Next delay
* `ts_pot`: Index of the time series used (in case of multivariate time series)
* `L_pot`: L statistic of next embedding step with delay `τ_pot` from `ts_pot`.
* `flag`: Did the embedding converge? i.e. L can not be further minimized anymore

MCDTS.nonlin_noise_reduction — Method

Simple nonlinear noise reduction algorithm by Schreiber 1993

Params:
*`m`           denotes the local embedding dimension
*`epsilon`     denotes the local neighborhood size as number of neighbours

Returns:
the filtered signal

Note that by applying this filter, there will be lost `m-1` datapoints.
We therefore phase-shift each datapoint in the resulting signal by `(m-1)/2`.

MCDTS.pecora — Method

pecora(s, τs, js; kwargs...) → ⟨ε★⟩, ⟨Γ⟩

Compute the (average) continuity statistic ⟨ε★⟩ and undersampling statistic ⟨Γ⟩ according to Pecora et al.[Pecoral2007] (A unified approach to attractor reconstruction), for a given input s (timeseries or Dataset) and input generalized embedding defined by (τs, js), according to genembed. The continuity statistic represents functional independence between the components of the existing embedding and one additional timeseries. The returned results are matrices with size TxJ.

Keyword arguments

delays = 0:50: Possible time delay values delays (in sampling time units). For each of the τ's in delays the continuity-statistic ⟨ε★⟩ gets computed. If undersampling = true (see further down), also the undersampling statistic ⟨Γ⟩ gets returned for all considered delay values.

J = 1:dimension(s): calculate for all timeseries indices in J. If input s is a timeseries, this is always just 1.

samplesize::Real = 0.1: determine the fraction of all phase space points (=length(s)) to be considered (fiducial points v) to average ε★ to produce ⟨ε★⟩, ⟨Γ⟩

K::Int = 13: the amount of nearest neighbors in the δ-ball (read algorithm description). Must be at least 8 (in order to gurantee a valid statistic). ⟨ε★⟩ is computed taking the minimum result over all k ∈ K.

metric = Chebyshev(): metrix with which to find nearest neigbhors in the input embedding (ℝᵈ space, d = length(τs)).

w = 1: Theiler window (neighbors in time with index w close to the point, that are excluded from being true neighbors). w=0 means to exclude only the point itself, and no temporal neighbors.

undersampling = false : whether to calculate the undersampling statistic or not (if not, zeros are returned for ⟨Γ⟩). Calculating ⟨Γ⟩ is thousands of times slower than ⟨ε★⟩.

db::Int = 100: Amount of bins used into calculating the histograms of each timeseries (for the undersampling statistic).

α::Real = 0.05: The significance level for obtaining the continuity statistic

p::Real = 0.5: The p-parameter for the binomial distribution used for the computation of the continuity statistic.

Description

Notice that the full algorithm is too large to discuss here, and is written in detail (several pages!) in the source code of pecora.

MCDTS.perform_recurrence_analysis — Method

Perform the RecurrenceAnalysis of some reconstruction trajectories `Y₁`, `Y₂`,
`Y₃`. Specifically, compute the fraction of recurrence rates from the
"original"/reference trajectory `Y_ref` with the one from the JRP of the
original `Y₁`, `Y₂`, `Y₃` together with the reconstructed trajectory. Also
compute RQA quantifiers of the recurrence plots of `Y₁`, `Y₂`, `Y₃` and `Y_ref`.

## Keyword arguments:
*`ε = 0.05`: The used threshold for constructing the recurrence plots
    The reconstruction method is fixed recurrence rate.
*`w = 1`: Theiler window used for all Datasets
*`lmin = 2`: Minimum used line length for digaonal line based RQA measures.
*`kNN = 1`: The number of nearest neighbors used for obtaining the mutual
    nearest neighbors measure

MCDTS.pick_possible_embedding_params — Method

Compute all possible τ-values (and according time series numbers) and their
corresponding Loss-statistics for the input delay_pre_selection_statistic `dps`.

MCDTS.rw_norm — Method

Compute the scaling term for the MASE measure. This is tge average in-sample
forecast error for a random-walk prediction, which uses the previous value in
the observed signal `x` as the forecast. `Tw` is the prediction time horizon.

MCDTS.softmaxL — Method

softmaxL(Ls; β=1.5)

Returns an index with prob computed by a softmax of all Ls.

MCDTS.uzal_cost_pecuzal_mcdts — Method

uzal_cost_pecuzal_mcdts(Y1::Dataset, Y2::Dataset, Tw; kwargs...) → L_decrease

This function is based on the functionality of [`uzal_cost`](@ref), here
specifically tailored for the needs in the PECUZAL algorithm.
Compute the L-statistics `L1` and `L2` for the input datasets `Y1` and `Y2` for
increasing time horizons `T = 1:Tw`. For each `T`, compute `L1` and `L2` and
decrease `L_decrease = L2 - L1`. If `L_decrease` is a negative value, then `Y2`
can be regarded as a "better" reconstruction that `Y1`. Break, when `L_decrease`
reaches the 1st local minima, since this will typically also be the global
minimum. Return the according minimum `L_decrease`-value.

## Keyword arguments

* `K = 3`: the amount of nearest neighbors considered, in order to compute σ_k^2
  (read algorithm description).
  If given a vector, minimum result over all `k ∈ K` is returned.
* `metric = Euclidean()`: metric used for finding nearest neigbhors in the input
  state space trajectory `Y.
* `w = 1`: Theiler window (neighbors in time with index `w` close to the point,
  that are excluded from being true neighbors). `w=0` means to exclude only the
  point itself, and no temporal neighbors.
* `econ::Bool = false`: Economy-mode for L-statistic computation. Instead of
  computing L-statistics for time horizons `2:Tw`, here we only compute them for
  `2:2:Tw`.
* `tws::Range = 2:Tw`: Further customization of the sampling of the different T's.
  While `econ=true` gives `tws = 2:2:Tw`, here any kind of interger ranges (starting at 2)
  are allowed, up to `Tw`.

MCDTS.zeroth_prediction_cost — Method

zeroth_prediction_cost(Y::Dataset; kwargs...) → Cost

Compute the mean squared one-step prediction error `Cost` of the Dataset `Y`.
The prediction is based on [`local_zeroth_prediction`](@ref).

## Keyword arguments

* `samplesize = 1.0`: Number of considered fiducial points v as a fraction of
  input state space trajectory `Y`'s length, in order to average the conditional
  variances and neighborhood sizes (read algorithm description) to produce `L`.
* `K = 3`: the amount of nearest neighbors considered, in order to compute the
  mean squared prediction error (read algorithm description).
* `metric = Euclidean()`: metric used for finding nearest neigbhors in the input
  state space trajectory `Y.
* `w = 1`: Theiler window (neighbors in time with index `w` close to the point,
  that are excluded from being true neighbors). `w=0` means to exclude only the
  point itself, and no temporal neighbors.
* `Tw = 1`: The time horizon for predictions. The `Cost` is the average error
  over these timesteps.

MCDTS.zeroth_prediction_cost_KL — Method

zeroth_prediction_cost_KL(Y::Dataset; kwargs...) → Cost

Compute the KL-divergence Cost of the Dataset Y. The prediction is based on local_zeroth_prediction.

Keyword arguments

samplesize = 1.0: Number of considered fiducial points v as a fraction of input state space trajectory Y's length, in order to average the conditional variances and neighborhood sizes (read algorithm description) to produce L.

K = 3: the amount of nearest neighbors considered, in order to compute the mean squared prediction error (read algorithm description).

metric = Euclidean(): metric used for finding nearest neigbhors in the input state space trajectory `Y.

w = 1: Theiler window (neighbors in time with index w close to the point, that are excluded from being true neighbors). w=0 means to exclude only the point itself, and no temporal neighbors.

Tw = 1: The time horizon for predictions. The Cost is the average error over these timesteps.

Pecora2007Pecora, L. M., Moniz, L., Nichols, J., & Carroll, T. L. (2007). A unified approach to attractor reconstruction. Chaos 17(1).